# Fix one of performance regressions in method calling