@Watson1978 2017-03-31
Looks like that a lot of `nop` instructions also are invoked around `ruby_th_dtrace_setup()` with clang's generated binary when run same test codes.
![](https://d3uepj124s5rcx.cloudfront.net/items/17023U0q290W1G0F1V1B/170331-0001.png?v=38690013)