@Watson1978 2017-04-04
Every time, Ruby checks whether DTrace's prove is enabled when it would call a method.
However, DTrace's prove would not provided with dynamically, it is detected to provide at only compilation time.

![](https://d3uepj124s5rcx.cloudfront.net/items/303c370C090f2v0t2V1V/170404-0001.png?v=ba555267)