@knu 2017-04-04
Just a minor nit, but I think you mean probe, not prove.