@Watson1978 2017-05-20
Looks for me that DTrace's probe will be enabled/disabled with `configure` script.

1. If the system has `dtrace` command, `configure` will mark DTrace is availed.
2. If DTrace is availed, DTrace's probe will be generated with `tool/gen_dummy_probes.rb` at `make` process.

And there is no way to enable/disable DTrace's probe at runtime.

Hmm, I wonder If I missed something.