@nurse 2017-05-20
There were some detailed documents about this mechanism before, but I couldn't find good document now. Therefore following are abstract of dtrace probes mechanism...

* Whether probes are inserted into the binary or not is statically decided.
  * It depends on the existence of `dtrace` command and configure as you wrote.
  * Those probe is disabled by default (exist in the binary but don't fire event; many `NOP`s you saw are disabled probes) 
* Whether probes fire the event is dynamically decided
  * Users can enable those probes by dtrace command. It dynamically replaces `NOP`s in the memory with the related probe instructions.