@Watson1978 2017-05-21
Thank you for replying.
It is very interesting mechanism and it sounds my changing is too bad.