@ko1 2017-05-21
It contains two modifications.
Could you separate them?

Method calling is common operation so that we need to care about them.
Please make your benchmarks more carefully.