@Watson1978 2017-05-22
```ruby
require 'benchmark/ips'

Benchmark.ips do |x|

  x.report "while" do |iter|
    ary = [1, 2, 3, 4, 5]
    i = 0
    while i < iter do
      ary.at(2)
      i += 1
    end
  end

  x.report "Integer#times" do |iter|
    ary = [1, 2, 3, 4, 5]
    iter.times do
      ary.at(2)
    end
  end

end
```