@Watson1978 2017-05-22
## Benchmark result
Using Apple LLVM version 8.1.0 (clang-802.0.41)

### Ruby 2.4.1
```
               while     20.988M (± 2.5%) i/s -    104.955M in   5.004269s
       Integer#times     14.062M (± 3.0%) i/s -     70.306M in   5.004882s
```

### Before (r58192)
```
               while     20.937M (± 2.5%) i/s -    104.668M in   5.002974s
       Integer#times     13.040M (± 1.9%) i/s -     65.238M in   5.005069s
```

* Result : ruby 2.4.1 vs r58192
```
while         -> 0.27 % down
Integer#times -> 7.76 % down
```

### After (r58192 + patches)
```
               while     21.409M (± 3.0%) i/s -    107.055M in   5.005548s
       Integer#times     13.159M (± 2.6%) i/s -     65.822M in   5.005854s
```

* Result : r58192 vs r58192 + patchs
```
while         -> 2.28 % up
Integer#times -> 0.90 % up
```