@Watson1978 2017-05-22
## Benchmark result
Using gcc-7 (Homebrew GCC 7.1.0) 7.1.0

### Ruby 2.4.1
```
               while     22.672M (± 2.3%) i/s -    113.292M in   5.000184s
             16.818M (± 1.9%) i/s -     84.169M in   5.006644s
```

### Before (r58192)
```
               while     22.439M (± 2.0%) i/s -    112.208M in   5.002965s
       Integer#times     16.613M (± 3.2%) i/s -     83.029M in   5.003500s
```

* Result : ruby 2.4.1 vs r58192
```
while         -> 0.97 % down
Integer#times -> 1.37 % down
```


### After (r58192 + patches)
```
               while     22.893M (± 1.7%) i/s -    114.492M in   5.002894s
       Integer#times     16.853M (± 1.8%) i/s -     84.302M in   5.004057s
```

* Result : r58192 vs r58192 + patchs
```
while         -> 2.04 % up
Integer#times -> 1.53 % up
```