@Watson1978 2017-05-22
Updated benchmark script and result in above.

I found,

1. the regression is occurred in clang + Integer#times combination.
2. The patches in this pull request increase performance very slightly.

So, I guess we need to focus on yield performance such as https://github.com/ruby/ruby/pull/1535