@nurse 2017-05-22
This improvement is small but this simplifies code.
(whether a patch should be merged or not depends the trade off between improvement and code complexity)
If this is also true on Linux+gcc, I'm positive to merge.
How do you think @ko1?