@ko1 2017-05-22
(1) Please explain your environment. You only show your compiler, but you need to show your CPU arch (and OS if it is matter). On this case, "LIKELY" affects branch prediction and CPU arch is a matter (of course, compiler version are also matter). dtrace (systemtap) enable/disable is also matter.
(2) Please show results in a table form. It is easy to compare.
(3) (advanced) Please write the reason why these fix are affect your benchmark results. I'm not sure why such performance changing. I need to understand "why" to approve this PR.
(4) Please explain why you change the benchmark. Benchmark should have the purpose and please make clear what is measured. Additionally, on this case, loop overhead is more matter so you need to measure only method call overhead. This is what I wrote " Please make your benchmarks more carefully." in above message.

Thanks,
Koichi