@ko1 2017-05-22
> I need to understand "why" to approve this PR.

In other words, I'll check the reason sometime if nobody explains.