@ko1 2017-05-22
I forgot one point.

(5) (related to (3)) It contains two modification. Please separate and show which is matter for performance. I'm not sure about dtrace, but method caching seems good point to use `LIKELY`.